+++ src/app/_layout.tsx
@@ -1,10 +1,9 @@
 import AppProviders from '@/components/providers'
-import { toastConfig } from '@/components/toast/ToastConfig'
 import useAppStore, { serializeCookieObject } from '@/hooks/stores/useAppStore'
 import { useModalStore } from '@/hooks/stores/useModalStore'
 import useCheckUpdate from '@/hooks/useCheckUpdate'
 import { initializeSentry } from '@/lib/config/sentry'
-import drizzleDb, { expoDb } from '@/lib/db/db'
+import drizzleDb from '@/lib/db/db'
 import lyricService from '@/lib/services/lyricService'
 import { ProjectScope } from '@/types/core/scope'
 import { toastAndLogError } from '@/utils/error-handling'
@@ -18,13 +17,12 @@ import { focusManager, onlineManager } from '@tanstack/react-query'
 import { useMigrations } from 'drizzle-orm/expo-sqlite/migrator'
 import * as Network from 'expo-network'
 import { SplashScreen, Stack, useNavigationContainerRef } from 'expo-router'
-import { useSQLiteDevTools } from 'expo-sqlite-devtools'
 import * as Updates from 'expo-updates'
 import { useCallback, useEffect, useState } from 'react'
 import type { AppStateStatus } from 'react-native'
 import { AppState, Platform, StyleSheet, View } from 'react-native'
 import { Text } from 'react-native-paper'
-import Toast from 'react-native-toast-message'
+import { Toaster } from 'sonner-native'
 import migrations from '../../drizzle/migrations'
 
 const logger = log.extend('UI.RootLayout')
@@ -54,9 +52,6 @@ export default Sentry.wrap(function RootLayout() {
 	useCheckUpdate()
 
 	useLogger(ref)
-
-	// eslint-disable-next-line @typescript-eslint/no-unsafe-call
-	useSQLiteDevTools(expoDb)
 
 	onlineManager.setEventListener((setOnline) => {
 		const eventSubscription = Network.addNetworkStateListener((state) => {
@@ -256,7 +251,7 @@ export default Sentry.wrap(function RootLayout() {
 					options={{ headerShown: false }}
 				/>
 			</Stack>
-			<Toast config={toastConfig} />
+			<Toaster />
 		</AppProviders>
 	)
 })